README.md and beam and main dir files are modified

+++ makedir.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 outdir=$1
+mode=$2 #0:keep existing dir 1:delete existing dir
+
 if [ -e $outdir ];then
-    rm -rf $outdir/*
+    if [ $mode -eq 1 ];then
+	rm -rf $outdir/*
+    fi
 else
-    mkdir $outdir
-fi 
+	mkdir $outdir
+fi
